fix: storage connection namespace in tests
